Add CloudWatch dashboard for head node metrics
When creating a cluster, a CloudWatch dashboard will be automatically created. It contains information about the Head Node: metrics about the EC2 instance, the EBS/RAID volumes and EFS/FSx file systems.
Concerning testing: one test is checking the jinja template rendering and another one is checking that the function is called when calling _create_bucket_with_resources. Manually tests were done with sge, torque and awsbatch but most of the tests were done with slurm using HIT clusters.
The code is flexible: it allows to add extra_parameters to properties of each widget (such as the title, as is already done). The problem with this extension is that some problems might happen when testing: if a sentence in the rendered template is too long, it will break the rendering_template test.
Open point: This commit enables the dashboard every time, the user can neither delete it using pcluster commands nor cancel the creation by having a specific parameter in the config file.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>