Add CloudWatch Logs Dashboard
Before this commit, a HeadNode dashboard was automatically created for the user when creating a cluster. Now, a logs dashboard is also created. This contains all the useful logs that are available with AWS ParallelCluster. The code is flexible and makes it possible to add easily other logs in the future but also makes it possible to choose the filters we apply to the log (only see warnings for example by modifying the regex).

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>